% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library.R
\name{getDistMat}
\alias{getDistMat}
\title{Obtains distance matrix}
\usage{
getDistMat(latlon)
}
\arguments{
\item{latlon}{An nx2 matrix of latitude and longitude points. See latlon().}
}
\description{
Given a nx2 matrix of latitude and longitude points, getDistMat will create an nxn matrix of distances between the points using Vincenty's formula.
}
\examples{
latlon <- getLatLon(c('New York, New York', 'Boston, MA'))
getDistMat(latlon)
}
\keyword{Distance}
\keyword{Matrix}

