% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library.R
\name{getGraphLines}
\alias{getGraphLines}
\title{Obtains minimum spannign tree edges for plotting.}
\usage{
getGraphLines(mst, latlon, N = 40, intermediate = TRUE)
}
\arguments{
\item{mst}{A minimum spanning tree graph object.}

\item{latlong}{The corresponding latlon points for the mst graph.}
}
\description{
Returns a list containing coordinates for plotting each edge of the minimum spanning tree using the greater circle path between the points.
}
\examples{
latlon <- getLatLon(c('New York, New York', 'Boston, MA'))
dist_mat <- getDistMat(latlon)
mst <- getMST(dist_mat)
getGraphLines(mst, latlon)
}
\keyword{minimum}
\keyword{mst,}
\keyword{plot}
\keyword{spanning}
\keyword{tree,}

