% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library.R
\name{getLatLon}
\alias{getLatLon}
\title{Obtains Latitutde and Longitude Coordinates.}
\usage{
getLatLon(cities, source = "google")
}
\arguments{
\item{cities}{A vector of geographic locations.}
}
\description{
Given a vector of geoegraphic locations, it will return a matrix of latitude and longitude points with rownames that take on the value of `cities`.
}
\examples{
getLatLon(c('New York, New York'))
}
\keyword{geocode}

