% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library.R
\name{getMST}
\alias{getMST}
\title{Obtains minimum spanning tree.}
\usage{
getMST(dist_mat)
}
\arguments{
\item{dist_mat}{A matrix of distances between points.}
}
\description{
Returns a graph object containing the minimum spanning tree between the points in dist_mat.
}
\examples{
latlon <- getLatLon(c('New York, New York', 'Boston, MA'))
dist_mat <- getDistMat(latlon)
getMST(dist_mat)
}
\keyword{minimum}
\keyword{mst,}
\keyword{spanning}
\keyword{tree}

