% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library.R
\name{graphLines2PolyLine}
\alias{graphLines2PolyLine}
\title{Converts list of mst edges to Polylines shapefile.}
\usage{
graphLines2PolyLine(graph_lines)
}
\arguments{
\item{graph_lines}{A list with each entry containing the coordinates of a minimum spanning tree's edge to plot}
}
\description{
Returns a Polylines shapefile. Allows for easier saving.
}
\examples{
latlon <- getLatLon(c('New York, New York', 'Boston, MA'))
dist_mat <- getDistMat(latlon)
mst <- getMST(dist_mat)
graph_lines <- getGraphLines(mst, latlon)
graphLines2Polyline(graph_lines)
}
\keyword{minimum}
\keyword{mst,}
\keyword{shapefile}
\keyword{spanning}
\keyword{tree,}

