% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library.R
\name{plotGraphLines}
\alias{plotGraphLines}
\title{Plots edges of minimum spanning tree.}
\usage{
plotGraphLines(graph_lines, line_col = "red", map = "county")
}
\arguments{
\item{graph_lines}{A list with each entry containing the coordinates of a minimum spanning tree's edge to plot}

\item{line_col}{The color the lines should be plotted. Defaults to red.}

\item{map}{The map from the `maps` package to draw as the background. Defaults to 'county'.}
}
\description{
Uses maps from the `maps` package as a background.
}
\examples{
latlon <- getLatLon(c('New York, New York', 'Boston, MA'))
dist_mat <- getDistMat(latlon)
mst <- getMST(dist_mat)
graph_lines <- getGraphLines(mst, latlon)
plotGraphLines(graph_lines)
}
\keyword{minimum}
\keyword{mst,}
\keyword{plot}
\keyword{spanning}
\keyword{tree,}

